// src/erp-ui/src/app/shared/models/permission-type.ts
export enum PermissionType {
  ItemTypesCanRead = 'itemtypes.canread',
  ItemTypesCanWrite = 'itemtypes.canwrite',
  ProductsCanRead = 'products.canread',
  ProductsCanWrite = 'products.canwrite',
  SuppliersCanRead = 'suppliers.canread',
  SuppliersCanWrite = 'suppliers.canwrite',
  UsersCanRead = 'users.canread',
  UsersCanWrite = 'users.canwrite',
  RolesCanRead = 'roles.canread',
  RolesCanWrite = 'roles.canwrite',
  PermissionsCanRead = 'permissions.canread',
  PermissionsCanWrite = 'permissions.canwrite',
}